+++ ChargeBee/Properties/AssemblyInfo.cs
@@ -29,8 +29,8 @@ using System.Runtime.InteropServices;
 // You can specify all the values or you can default the Build and Revision Numbers
 // by using the '*' as shown below:
 // [assembly: AssemblyVersion("1.0.*")]
-[assembly: AssemblyVersion("2.2.4")]
-[assembly: AssemblyFileVersion("2.2.4")]
-[assembly: AssemblyCompanyAttribute("kogir")]
+[assembly: AssemblyVersion("2.2.8")]
+[assembly: AssemblyFileVersion("2.2.8")]
+[assembly: AssemblyCompanyAttribute("Real Artists, Inc.")]
 [assembly: AssemblyProductAttribute("RealArtists.ChargeBee")]
 [assembly: AssemblyCopyrightAttribute("©ChargeBee Inc. and Nick Sivo. See LICENSE.")]